*)rename folder: "modules" to "common" *)rename file: "common.js" to "util.js" -)delete file: "tool.js"
