+++ common/pageInit.js
