+++ common/util.js
