--- modules/tool.js
@@ -1,14 +1,0 @@
-// var moment = require('moment');
-
-// moment.locate('zh-cn');
-
-// // 格式化时间
-// exports.formatDate = function(date, friendly) {
-// 	date = moment(date);
-
-// 	if (friendly) {
-// 		return date.fromNow();
-// 	} else {
-// 		return date.format('YYYY-MM-DD HH:mm');
-// 	}
-// };
